requirements.txt
psycopg2-binary==2.9.9
google-genai==0.3.0
python-dotenv==1.0.1